+++ CHANGELOG.md
@@ -3,13 +3,14 @@
 ## Version 4.0.6 (So far...) [Hero System 6e (Unofficial) v2](https://github.com/dmdorman/hero6e-foundryvtt)
 
 - Migrations are no longer supported from versions before 3.0.76. Migrate through any version prior to 4.0.6 if you find your world in this situation.
-- Initial support for Only In Heroic ID. There is a checkbox on the OTHER tab. Powers/skills with OIHID will be disabled when not in heroic ID. [#232](https://github.com/dmdorman/hero6e-foundryvtt/issues/232)
 - New effect icons for danger sense disabled, detect sense disabled, mental sense disabled, radio sense disabled, smell/taste sense disabled, sonar sense disabled, spatial awareness disabled, touch sense disabled. Abort icon corrected. [#1443](https://github.com/dmdorman/hero6e-foundryvtt/issues/1443)
 - Flight toggle now shows/hides the flight effect. [#1444](https://github.com/dmdorman/hero6e-foundryvtt/issues/1444)
 - Initial support for Only in Alternate Identity (OIAID). There is a "Heroic Identity" checkbox in the OTHERS tab. [#1431](https://github.com/dmdorman/hero6e-foundryvtt/issues/1431) [#232](https://github.com/dmdorman/hero6e-foundryvtt/issues/232)
 - Improved automatic spending of END during combat for continuous powers to prevent spending of resources when rewinding or re-doing a turn. [#1448](https://github.com/dmdorman/hero6e-foundryvtt/issues/1448)
 - MAX characteristic is now color coded when different than CORE. [#1461](https://github.com/dmdorman/hero6e-foundryvtt/issues/1461)
 - Improved display of COMPOUNDPOWER costs. [#1434](https://github.com/dmdorman/hero6e-foundryvtt/issues/1461)
+- FullHealth now also sets characteristic MAX to match CORE.
+- Added a dialog box for GM to select target list when confirming AOE placement on behalf of the player. [#1453](https://github.com/dmdorman/hero6e-foundryvtt/issues/1453)
 
 ## Version 4.0.5